Updated ls370 ipf for new RPi API
Instead of a token an auth code is sent as a second line of the header in get/post.
max_age_s has been removed from get/post because it doesn't do anything now, but I have left it everywhere else because it still needs to be implemented another way.