Added delay, fixed initfastdac()
added optional delay to fdacrecordvalues - Since fdacrecord values ramps to beginning of 1D sweep, better to have delay passed in there so that it will definitely wait for specified time after ramping to beginning of scan (e.g. in 2D sweep).

Initfastdac had a small mistake, and then I just added some description of the dac and adc waves.
Changed ADC input to default to "" instead of "0" since 0 is misleading.