Added verbose option to fetchData, added SQL_format_time, added sql_response_code
verbose option is verbose by default. added so that a request for data which may be missing won't always print a warning message.
added sql_repsonse_code global variable so that it is easy to check if there was a warning to be handled somewhere else.
Added SQL_format_time to convert from seconds since 1904 to the SQL TIMESTAMP format