Changed to make closer to working with current setup
Changed to look for LoggingSchedules.txt for now since that is what I have. Should be easy to change it back later.
Also made it use the channel lookups and look for data in the database by channel label rather than ID. Should make it work well with both LD and XLD system

Commented out heater stuff as it was just hard coded! This bit needs writing!