Added additional sanity check to SFfd_check_lims()
My problem yesterday was because I was making a channels list using igors addlistitem BUT Igor does not make lists correctly (adds separator to the end) which means when I concatenate the channelsx and channelsy list, the list gets a double separator and was breaking the check_lims loop which looks at the number of items in the list. Added check for double separator.