Updates to BabyDAC
Can ramp with label name or channel number now

Get status now returns the values stored in old_dacvalstr (which is more likely to be where the DACs actually are. e.g. if typed in new value but not ramped, dacvalstr will not match the true DAC outputs, but old_dacvalstr will).