FastDAC updates
Fixed getstatus to not put quotes around numbers

Made clearbuffer show what it is clearing, useful for debugging too

Improved spectrum analyser window plotting so that it's easier to add lines to a graph to compare them. It won't keep creating new windows now.

Spectrum analyser uses Periodogram with parsevals now to make power spectrum. This seems to make the most sense after working with Josn on it.