Updating spectrum_analyzer to be faster
spectrum_analyzer uses the built in  DSPPeriodogram to calculate the periodogram. This process works well when number of points in array are multiples of 2. e.g. 2,4,8,16,32... So here we slice the array so that we have the maximum number of points that is a multiple of 2.
this will greatly spped the calculation but worst  case we may could only be calculating the periodogram on half of the data.