updated how the waves are initialized
the waves had to pass a condition that looked at the value of a number indicated in the wave name. The default wave name for example would be wave0. It used the "0" value to pass the condition on line 963. It previously did not account for name changes (we often change the names, for example to "cscurrent". To work around this issue, I am using the raw wave names (ADC0,ADC1.,....) to pass the condition