initializing and recording 1d and 2d demodulation waves
following the workflow in scanfastdac, We first initialize the waves for demodulation 1d and 2d as needed. These are kept separate from the original raw waves and calc waves. Then the same workflow in process and distribute where the 1d demod waves are copied into 2d as needed. there are nested if functions, specifically for the demodulation in y. If  "Save Demod.y"  box is checked in scancontroller, It will initialize and record a corresponding 2d wave.