scc_checkDeviceNumber function
this function will return the instrument IDs of all the channels being ramped and all the channels being recorded (if the optional adc = 1 parameter is passed). This will be useful to keep track of what to set to master/slave and the use of fake ramps/ recordings. Might be worthwhile picking a specific format and putting it into scanvars?