fit_entropy_util cleaned and basic working state
center_demod can take in an entropy dataset with repeats and 1) Center each row basedo n the cold traces. 2) Calculate the scaling factor, 3) Average and plot the demodulated entropy version and the integrated entropy version.
Next steps are to handle bid 2d datasets where we do not want to average.