added scanvars for adclists
have to adjust the comparison between fake and real recordings. The current method uses adclists to distribute data, it might be useful to have a kwListStr for the fdIDs pointing towards the adcs it recorded and the extra added for sync purposes. note.* adclists uses the numbers in the windows to target the correct graphs while my kwliststrs would contain the actual adc channels. might be useful to keep it as the same numbers in the windows? not entirely sure just yet