the issue was related to overwriting in the resampling process.
the issue was related to overwriting the resampling after doing the measurement. Therefore, we added an optional parameter called skip_raw2calc to the scfd_RecordValues() function which means that if it's set to 1, it will skip the reassignment of the calc waves based on raw waves.  So This parameter allows us to pass through 'scfd_raw2calcQuickDistribute()' in the scfd_RecordBuffer() function if it is not specified, and vice versa.